Refactor to library and add to lint.
