--- cmd/genesis/kcoin/vmtracer.go
@@ -1,39 +1,0 @@
-package kcoin
-
-import (
-	"github.com/kowala-tech/kcoin/common"
-	"github.com/kowala-tech/kcoin/core/vm"
-	"time"
-)
-
-//TODO:This vmTracer is an exact copy of the one in pupphet, can be unified? Or maybe it is not even needed? Someone
-//with more knowledge on this part is welcome.
-type vmTracer struct {
-	data map[common.Address]map[common.Hash]common.Hash
-}
-
-func newVmTracer() *vmTracer {
-	return &vmTracer{
-		data: make(map[common.Address]map[common.Hash]common.Hash, 1024),
-	}
-}
-
-func (vmt *vmTracer) CaptureState(env *vm.EVM, pc uint64, op vm.OpCode, gas, cost uint64, memory *vm.Memory, stack *vm.Stack, contract *vm.Contract, depth int, err error) error {
-	if err != nil {
-		return err
-	}
-	if op == vm.SSTORE {
-		s := stack.Data()
-		addrStorage, ok := vmt.data[contract.Address()]
-		if !ok {
-			addrStorage = make(map[common.Hash]common.Hash, 1024)
-			vmt.data[contract.Address()] = addrStorage
-		}
-		addrStorage[common.BigToHash(s[len(s)-1])] = common.BigToHash(s[len(s)-2])
-	}
-	return nil
-}
-
-func (vmt *vmTracer) CaptureEnd(output []byte, gasUsed uint64, t time.Duration, err error) error {
-	return nil
-}
